requirements.txt
streamlit
qrcode[pil]
pillow
reportlab
requests
